+++ konva.js
@@ -3,7 +3,7 @@
  * Konva JavaScript Framework v1.2.2
  * http://konvajs.github.io/
  * Licensed under the MIT or GPL Version 2 licenses.
- * Date: Tue Sep 20 2016
+ * Date: Wed Sep 21 2016
  *
  * Original work Copyright (C) 2011 - 2013 by Eric Rowell (KineticJS)
  * Modified work Copyright (C) 2014 - 2015 by Anton Lavrenov (Konva)
@@ -9057,7 +9057,7 @@
                 for(n = 0; n < len; n++) {
                     layer = layers[n];
                     layer.setSize(width, height);
-                    layer.draw();
+                    layer.batchDraw();
                 }
             }
         },