+++ src/Stage.js
@@ -346,7 +346,7 @@
                 for(n = 0; n < len; n++) {
                     layer = layers[n];
                     layer.setSize(width, height);
-                    layer.draw();
+                    layer.batchDraw();
                 }
             }
         },